Improve output for dita maps

+++ DitaDotNetLib/DitaConverter.cs
@@ -7,11 +7,11 @@ namespace DitaDotNet {
         // The collection of DITA content to convert
         protected DitaCollection Collection { get; set; }
 
-        // The bookmap that is the root of the collection
-        protected DitaBookMap BookMap { get; set; }
+        // The bookmap/amp that is the root of the collection
+        protected DitaFile RootMap { get; set; }
 
         // The main conversion action
-        public bool Convert(string input, string output, bool rename = false, bool deleteExistingOutput = false) {
+        public bool Convert(string input, string output, string rootMapFile, bool rename = false, bool deleteExistingOutput = false) {
             try {
                 // Make sure the output path exists
                 VerifyOutputPath(output);
@@ -20,14 +20,8 @@ namespace DitaDotNet {
                 Collection = new DitaCollection();
                 Collection.LoadDirectory(input);
 
-                // Is there a bookmap?
-                List<DitaBookMap> bookMaps = Collection.GetBookMaps();
-                if (bookMaps.Count == 1) {
-                    BookMap = bookMaps[0];
-                }
-                else {
-                    throw new Exception($"Expecting exactly 1 bookmap, but found {bookMaps.Count}");
-                }
+                // Find the rootmap
+                FindRootMap(rootMapFile);
 
                 // Try renaming the files in the collection, if requested
                 if (rename) {
@@ -49,24 +43,56 @@ namespace DitaDotNet {
         }
 
         // Validate that the output folder exists, and create it if it doesn't
-        protected void VerifyOutputPath(string output) {
+        private void VerifyOutputPath(string output) {
             if (!Directory.Exists(output)) {
                 Directory.CreateDirectory(output);
             }
         }
 
-
         // Delete existing output files
         protected void DeleteOutputFiles(string output) {
             try {
-                string[] files = Directory.GetFiles(output);
-                foreach (string file in files) {
-                    File.Delete(file);
-                    Trace.TraceInformation($"Deleted output file {file}.");
+                if (Directory.Exists(output)) {
+                    string[] files = Directory.GetFiles(output);
+                    foreach (string file in files) {
+                        File.Delete(file);
+                        Trace.TraceInformation($"Deleted output file {file}.");
+                    }
                 }
             }
-            catch {
-                Trace.TraceError($"Error deleting out files in {output}");
+            catch (Exception ex) {
+                Trace.TraceError($"Error deleting output files in {output}");
+                Trace.TraceError(ex);
+            }
+        }
+
+        // Try to find the root map file
+        private void FindRootMap(string rootMapFile) {
+            // If no file name specified, try to find a bookmap
+            if (string.IsNullOrEmpty(rootMapFile)) {
+                List<DitaBookMap> bookMaps = Collection.GetBookMaps();
+                if (bookMaps.Count == 1) {
+                    RootMap = bookMaps[0];
+                    return;
+                }
+                throw new Exception($"Expecting exactly 1 bookmap, but found {bookMaps.Count}");
+            }
+
+            // Is there a bookmap or map with the given name?
+            DitaFile rootFile = Collection.GetFileByName(rootMapFile);
+            if (rootFile == null) {
+                throw new Exception($"Specified root map file {rootMapFile} was not found in collection.");
+            }
+
+            switch (rootFile) {
+                case DitaBookMap bookMap:
+                    RootMap = bookMap;
+                    break;
+                case DitaMap map:
+                    RootMap = map;
+                    break;
+                default:
+                    throw new Exception($"{rootMapFile} must be a map or bookmap.");
             }
         }
     }
